sources/backend-v2/requirements.txt
flask
flask-mysqldb
flask-jwt-extended
flask-cors
python-dotenv
werkzeug
tensorflow
pillow
numpy
requests
joblib
scikit-learn
pandas
